+++ src/main/java/com/mycompany/traintrack/server/app/TestsForGitub/testForBranchDev.java
@@ -1,4 +1,4 @@
-package com.mycompany.traintrack.server.app;
+package com.mycompany.traintrack.server.app.TestsForGitub;
 
 public class testForBranchDev {
     public testForBranchDev(){
