+++ src/main/java/com/mycompany/traintrack/server/app/TestsForGitub/testServer3.java
@@ -1,4 +1,4 @@
-package com.mycompany.traintrack.server.app;
+package com.mycompany.traintrack.server.app.TestsForGitub;
 
 public class testServer3 {
     public testServer3(){
