--- src/main/java/com/mycompany/traintrack/server/app/hola.java
@@ -1,5 +1,0 @@
-package com.mycompany.traintrack.server.app;
-
-public class hola {
-    
-}
